+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.3.0
+
+* Rename `Zonex.list/2` to `Zonex.list_canonical/2`
+* Rename `Zonex.get/3` to `Zonex.get_canonical/3`
+* Rename `Zonex.get!/3` to `Zonex.get_canonical!/3`
+* Rename Zone `standard` property to `golden`
+
 ## 0.2.0
 
 * Add GitHub source URL in mix project config